# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurnByTurn API",
    "version": "1.0.0",
    "description": "Backend for TurnByTurn, a rotating savings (ajo / esusu) app. All money values are integer kobo. Errors always return `{ error: { code, message, details?, retryAfter? } }`; see docs/ERRORS.md for the code → screen map."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "VALIDATION_ERROR",
                  "NOT_FOUND",
                  "CONFLICT",
                  "INTERNAL_ERROR",
                  "GATEWAY_ERROR",
                  "INVALID_CREDENTIALS",
                  "OTP_INVALID",
                  "OTP_EXPIRED",
                  "OTP_RESEND_COOLDOWN",
                  "PASSWORD_POLICY",
                  "ACCOUNT_EXISTS",
                  "TOO_MANY_ATTEMPTS",
                  "BANK_VERIFICATION_FAILED",
                  "SESSION_EXPIRED",
                  "UNAUTHENTICATED",
                  "ACCESS_DENIED",
                  "SIGNUP_INCOMPLETE",
                  "GROUP_NOT_DRAFT",
                  "GROUP_NOT_ACTIVE",
                  "GROUP_SETUP_INCOMPLETE",
                  "GROUP_FULL",
                  "INVITE_INVALID",
                  "ALREADY_MEMBER",
                  "AMOUNT_INVALID",
                  "NOTHING_DUE",
                  "CYCLE_CLOSED",
                  "PAYOUT_NOT_ELIGIBLE",
                  "PAYOUT_ACCOUNT_MISSING",
                  "REMINDER_COOLDOWN"
                ]
              },
              "message": {
                "type": "string"
              },
              "retryAfter": {
                "type": "integer",
                "description": "Seconds; also sent as Retry-After header"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "PaymentAttempt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-f0-9]{24}$"
          },
          "reference": {
            "type": "string",
            "example": "TRX-0926-0042"
          },
          "kind": {
            "type": "string",
            "enum": [
              "full",
              "partial"
            ]
          },
          "amount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "serviceFee": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "totalCharged": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "success",
              "failed",
              "abandoned"
            ]
          },
          "checkoutUrl": {
            "type": "string"
          },
          "gatewayAmountReceived": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "amountCredited": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "excessAmount": {
            "type": "integer",
            "description": "Amount in kobo (₦1 = 100)"
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/dev/checkout/{reference}": {
      "get": {
        "tags": [
          "Dev (non-production)"
        ],
        "summary": "Mock Squadco checkout page",
        "operationId": "getDevCheckoutReference",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/dev/payments/{reference}/complete": {
      "post": {
        "tags": [
          "Dev (non-production)"
        ],
        "summary": "Simulate gateway confirmation (?fail=1, or body.amount to overpay)",
        "operationId": "postDevPaymentsReferenceComplete",
        "parameters": [
          {
            "name": "reference",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payment": {
                      "$ref": "#/components/schemas/PaymentAttempt"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "integer",
                    "description": "Amount in kobo (₦1 = 100)"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/dev/jobs/run": {
      "post": {
        "tags": [
          "Dev (non-production)"
        ],
        "summary": "Run scheduled jobs now",
        "operationId": "postDevJobsRun",
        "security": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}